Moved around some plugins
