+++ src/BlazorGPT.Web/Plugins/SamplesFun/Excuses/config.json
