+++ src/BlazorGPT.Web/Plugins/SamplesFun/Excuses/skprompt.txt
