+++ src/BlazorGPT.Web/Plugins/SamplesFun/Joke/skprompt.txt
