+++ src/BlazorGPT.Web/Plugins/SamplesFun/Limerick/config.json
