+++ src/BlazorGPT.Web/Plugins/SamplesFun/Limerick/skprompt.txt
